- added codable RecognitionResults struct
